+++ apps/service/src/app/report/[publishId]/[problemId]/layout.tsx
@@ -1,4 +1,4 @@
-import { ReportProvider } from '@contexts/ReportContext';
+import { ReportProvider } from '@contexts';
 
 const ReportLayout = ({
   children,
